requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
einops==0.6.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.3.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.1
lpips==0.1.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-service==2.4.0
monai==1.1.0
mpmath==1.2.1
multidict==6.0.4
oauthlib==3.2.2
Pillow==9.4.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
pytorch-lightning==1.6.4
pytorch-msssim==0.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==19.0.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tifffile==2023.2.28
torch==2.1.0.dev20230305
torchaudio==2.0.0.dev20230305
torchmetrics==0.11.3
torchvision==0.15.0.dev20230305
tqdm==4.65.0
triton==2.0.0
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
